@@ -1,48 +1,36 @@
 {
   "name": "kuberos",
-  "description": "An OIDC authentication helper for Kubernetes' kubectl",
   "version": "1.0.0",
-  "author": "Nic Cope <n+kuberos@rk0n.org>",
-  "license": "Apache-2.0",
-  "private": true,
+  "description": "An OIDC authentication helper for Kubernetes' kubectl",
+  "main": "webpack.config.js",
+  "author": "platforms@digital.justice.gov.uk",
+  "license": "MIT",
   "scripts": {
-    "dev": "cross-env NODE_ENV=development webpack serve --mode=development",
-    "build": "cross-env NODE_ENV=production webpack --progress --mode=production"
+    "build": "webpack --config webpack.config.js --mode=production --progress",
+    "dev": "webpack-dev-server --config webpack.config.js --hot --mode=development"
   },
+  "keywords": [],
   "dependencies": {
-    "@babel/plugin-transform-runtime": "^7.14.2",
-    "axios": "^0.21.2",
-    "cross-env": "^7.0.3",
-    "css-loader": "^5.2.4",
-    "element-ui": "^2.15.1",
-    "file-loader": "^6.2.0",
-    "highlight.js": ">=10.7.2",
-    "jquery": "^3.6.0",
-    "node-sass": "^7.0.0",
-    "sass-loader": "^12.6.0",
-    "style-loader": "^2.0.0",
-    "url-loader": "^4.1.1",
-    "vue": "^2.6.12",
-    "vue-axios": "^2.1.5",
-    "vue-highlightjs": "^1.3.3",
-    "vue-meta": "^1.4.0"
+    "@highlightjs/vue-plugin": "^2.1.0",
+    "axios": "^0.26.0",
+    "element-plus": "^2.0.2",
+    "element-theme-chalk": "^2.15.6",
+    "vue": "^3.2.31",
+    "vue-axios": "^3.4.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.2",
-    "@babel/preset-env": "^7.14.2",
-    "babel-loader": "^8.2.2",
-    "css-loader": "^5.2.4",
-    "element-theme-chalk": "^2.15.1",
-    "html-webpack-plugin": "^5.3.1",
-    "rimraf": "^3.0.2",
-    "sass-loader": "^12.6.0",
-    "style-loader": "^2.0.0",
-    "vue": "^2.6.12",
-    "vue-loader": "^15.9.7",
+    "@babel/core": "^7.17.5",
+    "@babel/preset-env": "^7.16.11",
+    "babel-loader": "^8.2.3",
+    "css-loader": "^6.6.0",
+    "html-webpack-plugin": "^5.5.0",
+    "jquery": "^3.6.0",
+    "url-loader": "^4.1.1",
+    "vue-loader": "^17.0.0",
     "vue-style-loader": "^4.1.3",
-    "vue-template-compiler": "^2.6.12",
-    "webpack": "^5.37.0",
-    "webpack-cli": "^4.7.0",
-    "webpack-dev-server": "^3.11.2"
+    "vue-template-compiler": "^2.6.14",
+    "webpack": "^5.69.1",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.7.4"
   }
 }
