@@ -1,14 +1,12 @@
-// Webpack uses this to work with directories
 var path = require('path')
-
 var webpack = require('webpack')
-const VueLoaderPlugin = require('vue-loader/lib/plugin');
+const { VueLoaderPlugin } = require('vue-loader')
+var mode = process.env.NODE_ENV || 'development';
 
-// This is the main configuration object.
-// Here, you write different options and tell Webpack what to do
 module.exports = {
-  // Path to your entry point. From this file Webpack will begin its work
-  entry: './src/main.js',
+  entry: [
+    './src/main.js'
+  ],
   // Path and filename of your result bundle.
   // Webpack will bundle all JavaScript into this file
   output: {
@@ -16,11 +14,10 @@ module.exports = {
     publicPath: '/dist/',
     filename: 'build.js'
   },
-  // You say Webpack how exactly it should transform different types of files
   module: {
     rules: [{
         test: /\.vue$/,
-        use: ['vue-loader']
+        use: 'vue-loader'
       },
       {
         test: /\.js$/,
@@ -29,7 +26,7 @@ module.exports = {
       },
       {
         test: /\.css$/,
-        use: ['style-loader', 'css-loader']
+        use: ['vue-style-loader', 'css-loader']
       },
       {
         test: /\.(png|jpg|jpeg|gif|eot|ttf|woff|woff2|svg|svgz)(\?.+)?$/,
@@ -39,29 +36,29 @@ module.exports = {
             limit: 10000
           }
         }]
-      }
+      },
+      {
+        test: /\.mjs$/i,
+        resolve: { byDependency: { esm: { fullySpecified: false } } },
+      },
     ]
+  }, 
+  devServer: {
+    static: __dirname
+  },  
+  devtool: (mode === 'development') ? 'inline-source-map' : false,
+  mode: mode,
+  performance: {
+    hints: false
   },
-  //Automatically load modules instead of having to import or require them everywhere.
   plugins: [
     new webpack.ProvidePlugin({
       '$': 'jquery'
     }),
     new VueLoaderPlugin(),
-  ],
-  resolve: {
-    alias: {
-      'jquery': "jquery/dist/jquery.min.js",
-      'vue$': 'vue/dist/vue.esm.js'
-    },
-    extensions: ['*', '.js', '.vue', '.json']
-  },
-  performance: {
-    hints: false
-  }
+  ]
 }
-// Additional plugins for production mode for minimize the code
-if (process.env.NODE_ENV === 'production') {
+if (mode === 'production') {
   // http://vue-loader.vuejs.org/en/workflow/production.html
   module.exports.plugins = (module.exports.plugins || []).concat([
     new webpack.DefinePlugin({
